fix interface type

+++ packages/contracts/contracts/interfaces/v0.8.x/ISharedMinterRAMV0.sol
@@ -11,6 +11,6 @@ interface ISharedMinterRAMV0 {
     function createBid(
         uint256 projectId,
         address coreContract,
-        uint8 slotIndex
+        uint16 slotIndex
     ) external payable;
 }
